Add files via upload
In this work, we focus on explanations for items that do not appear in the recommendations they way we expect them to, expressed in why-not questions, to aid the system
engineer improve the recommender.
That is, instead of offering explanations on every item proposed by the system, we allow the developer give feedback about items that were not proposed. We consider here the most traditional category of recommenders, i.e., the collaborative filtering.